Add timeout, remove handlers functionality
Re-add timeout functionality for urllib2 OpenDirector object calls
open() and remove handlers, because it's unneeded functionality for
the main project.

We use the timeout parameter to set a timeframe after which a
socket.timeout exception is raised by urllib2 or _ssl libraries.
This functionality is needed for the main project.